requirements.txt
requests==2.31.0
pandas==2.1.4
transformers==4.25.1 # Never upgrade
numpy==1.26.2
torch==2.1.2
scikit-learn==1.3.2
accelerate==0.25.0
xformers==0.0.23
evaluate==0.4.1
# mlflow==2.9.2
# huggingface_hub==0.18.0
tqdm==4.66.1
# spacy[apple]==3.7.2  # for mac
spacy==3.4.4  #
matplotlib==3.8.2
google-cloud-firestore==2.14.0
oauth2client==4.1.3
gspread==5.12.3
gspread_dataframe==3.3.1
scipy==1.11.4
# pycontractions==2.0.1
# spacy_transformers==1.3.3

# paper c ---------
setfit==1.0.1
sentence-transformers==2.2.2
datasets==2.15.0
openai==0.10.0
# pyarrow==5.0.0
afinn==0.1
optuna==3.5.0
unidecode==1.3.8
textblob== 0.17.1
seaborn==0.13.2
networkx==3.2.1
bertviz==1.4.0
lime==0.2.0.1
shap==0.44.1
# google-api-core==0.5.1
captum==0.7.0
joblib==1.3.2
statsmodels==0.14.1
datasets==2.15.0
contractions==0.1.73
transformers_interpret==0.10.0
